Implementación de CUD en un solo contenedor
