+++ nginx/default.conf
@@ -1,9 +1,6 @@
 # Define upstream servers para cada servicio
-upstream crear-personas {
-    server crear-personas:3001;
-}
-upstream modificar-personas {
-    server modificar-personas:3002;
+upstream personas-crud {
+    server personas-crud:3001;
 }
 upstream consultar-personas {
     server consultar-personas:3003;
@@ -11,9 +8,7 @@ upstream consultar-personas {
 # upstream consultar-nl-personas {
 #     server consultar-nl-personas:3004;
 # }
-upstream borrar-personas {
-    server borrar-personas:3005;
-}
+
 upstream logs-personas {
     server logs-personas:3006;
 }
@@ -21,6 +16,9 @@ upstream logs-personas {
 server {
     listen 80;
     server_name localhost;
+    
+    # Aumentar el tamaño máximo del cuerpo de la solicitud a 5MB
+    client_max_body_size 5M;
 
     # Frontend
     location / {
@@ -31,19 +29,16 @@ server {
 
     # Crear personas
     location /api/personas/crear {
-        # CORS headers
         add_header 'Access-Control-Allow-Origin' '*' always;
         add_header 'Access-Control-Allow-Methods' 'POST, OPTIONS' always;
         add_header 'Access-Control-Allow-Headers' 'Origin, Content-Type, Accept, Authorization' always;
 
-        # OPTIONS handling
         if ($request_method = 'OPTIONS') {
             return 204;
         }
 
-        # Rewrite URL
-        rewrite ^/api/personas/crear(.*)$ /$1 break;
-        proxy_pass http://crear-personas/;
+        rewrite ^/api/personas/crear(.*)$ /crear$1 break;
+        proxy_pass http://personas-crud/;
         proxy_http_version 1.1;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
@@ -59,8 +54,8 @@ server {
             return 204;
         }
 
-        rewrite ^/api/personas/modificar(.*)$ /$1 break;
-        proxy_pass http://modificar-personas/;
+        rewrite ^/api/personas/modificar/(.*)$ /modificar/$1 break;
+        proxy_pass http://personas-crud/;
         proxy_http_version 1.1;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
@@ -98,8 +93,8 @@ server {
             return 204;
         }
 
-        rewrite ^/api/personas/borrar(.*)$ /$1 break;
-        proxy_pass http://borrar-personas/;
+        rewrite ^/api/personas/borrar(.*)$ /borrar$1 break;
+        proxy_pass http://personas-crud/;
         proxy_http_version 1.1;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
