--- services/modificar-personas/dockerfile
@@ -1,7 +1,0 @@
-FROM node:18-alpine
-WORKDIR /app
-COPY package*.json ./
-RUN npm install
-COPY . .
-EXPOSE 3002
-CMD ["node", "src/index.js"]
