--- services/modificar-personas/package.json
@@ -1,13 +1,0 @@
-{
-    "name": "modificar-personas",
-    "version": "1.0.0",
-    "description": "Microservicio para modificar personas",
-    "main": "src/index.js",
-    "scripts": {
-      "start": "node src/index.js"
-    },
-    "dependencies": {
-      "express": "^4.18.2",
-      "pg": "^8.11.3"
-    }
-  }
