--- services/modificar-personas/src/db.js
@@ -1,13 +1,0 @@
-const { Pool } = require('pg');
-
-const pool = new Pool({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  port: 5432
-});
-
-module.exports = {
-  query: (text, params) => pool.query(text, params)
-};
