+++ services/personas-crud/src/db.js
@@ -1,9 +1,9 @@
 const { Pool } = require('pg');
 
 const pool = new Pool({
+  host: process.env.DB_HOST || 'postgres-db',
   user: process.env.DB_USER || 'admin',
   password: process.env.DB_PASSWORD || 'admin',
-  host: process.env.DB_HOST || 'postgres-db',
   database: process.env.DB_NAME || 'gestion_datos',
   port: process.env.DB_PORT || 5432,
 });
