--- services/personas/Dockerfile
@@ -1,9 +1,0 @@
-FROM node:18-alpine AS build
-
-WORKDIR /app
-COPY package*.json ./
-RUN npm install
-COPY . .
-EXPOSE 3001
-
-CMD ["npm", "run", "start:dev"]
