--- services/personas/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "personas",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node src/index.js",
-    "start:dev": "nodemon src/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "commonjs",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^5.1.0",
-    "pg": "^8.11.3",
-    "pgvector": "^0.1.8"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.9"
-  }
-}
